+++ src/components/footer/footer.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+class Footer extends React.Component {
+  render() {
+    return <footer>&copy; 2018 Code Fellows</footer>;
+  }
+}
+
+export default Footer;
